requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-openai
langchain-huggingface
langchain_community
sentence-transformers
python-multipart
faiss-cpu
loguru